IOS-4565 Moved validator to its own file
